% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{addFilter}
\alias{addFilter}
\title{Add filter to bsj.read-data.}
\usage{
addFilter(object, filter, mode = "strict", ...)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{filter}{A logical vector (circSample) or list of logical vectors (circExperiment) indicating which bsj.reads should be included in downstream analyses.}

\item{mode}{Character, "strict" sets new filter value to FALSE if either the previous \cite{include.read} or new filter value is \code{FALSE}. "last" overwrites values with the new regardles of previous value.}
}
\value{
Object with same class as input.
}
\description{
Add filter to bsj.read-data.
}
\examples{
#addFilter(object, filter)

}
