% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{addLinCounts}
\alias{addLinCounts}
\title{Add information on usage of donor and acceptor sites for linear splicing}
\usage{
addLinCounts(cbj = cbj.sum, sj = sj.data)
}
\description{
Add information on usage of donor and acceptor sites for linear splicing
}
