% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{adjustAlignment}
\alias{adjustAlignment}
\title{Read alignment to nearest genes}
\usage{
adjustAlignment(object, cores = 1L, ...)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{cores}{Only supported on unix! Number of cores used for parallel processing for circExperiment objets.}

\item{...}{}
}
\value{
A new object with shifted bsj.read data
}
\description{
Reads circSample bsj.reads, shifts shiftable reads, and replaces bsj.read data with the shifted data
}
\examples{
#adjustAlignment(object)


}
