% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{alignmentStats}
\alias{alignmentStats}
\title{Extract alignment stats}
\usage{
alignmentStats(object, out_type = "wide")
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Datatable with STAR log file information
}
\description{
Parse STAR alignment log files, and extract alignment stats.
}
\examples{
#alignmentStats(object)

}
