% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNA_annotation.r
\name{annotateByOverlap}
\alias{annotateByOverlap}
\title{Annotate circRNAs (bsID) by identifying overlapping genes}
\usage{
annotateByOverlap(bsids = NULL, db = ahdb, t = "within")
}
\arguments{
\item{t}{Type of overlap}

\item{bsid}{Vector of bsIDs to annotate.}

\item{TXDB}{Annotation database (from AnnotationHub!!!)}
}
\description{
Annotate circRNAs (bsID) by identifying overlapping genes
}
