% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{bam.file}
\alias{bam.file}
\title{Accessor for the BAM file path}
\usage{
bam.file(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with the chimeric file path(s)
}
\description{
circSample and circExperiment class accessors for showing the chimeric file path(s) of circSample object(s)
}
\examples{
# The BAM file path of a sample object
bam.file(sample.object)

# The BAM file path of each sample in each sample of an experiment
bam.file(experiment.object)

}
