% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{bsid2junc}
\alias{bsid2junc}
\title{Convert bsID to GRanges object of acceptor or donor positions}
\usage{
bsid2junc(bsids = NULL, junc.type = "acceptor")
}
\description{
Convert bsID to GRanges object of acceptor or donor positions
}
