% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{bsj.counts<-}
\alias{bsj.counts<-}
\title{Replacement method for the backsplice junction count data}
\usage{
bsj.counts(object) <- value
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
New class object with replaced backsplice junction count data
}
\description{
circSample and circExperiment class replacement methods for replacing the backsplice junction count data
The replacment value must be valid to enable replacement
}
\examples{
# A datatable of the backsplice junction count data in a sample object
#bsj.counts(sample.object) <- new.datatable

# A datatable of the backsplice junction count data in each sample of an experiment
#bsj.counts(experiment.object) <- new.datatable.list

}
