% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{bsj.counts}
\alias{bsj.counts}
\title{Accessor for the backsplice junction count data}
\usage{
bsj.counts(object, returnAs = "list", ...)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{returnAs}{Determines if output should be list (default) or GRanges (GRangesList for circExperiment objects)}

\item{use.names}{Boolean. If TRUE, the list being returned will be named using sample IDs}
}
\value{
List of datatable(s) with backsplice junction count data
}
\description{
circSample and circExperiment class accessors for showing the backsplice junction count data
}
\examples{
# A datatable of the backsplice junction count data in a sample object
#bsj.counts(sample.object)

# A datatable of the backsplice junction count data in each sample of an experiment
#bsj.counts(experiment.object)

}
