% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{bsj.reads<-}
\alias{bsj.reads<-}
\title{Replacement method for the backsplice junction read data}
\usage{
bsj.reads(object) <- value
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
New class object with replaced filtered backsplice junction read data
}
\description{
circSample and circExperiment class replacement methods for replacing the filtered backsplice junction read data of a circSample object(s)
The replacment value must be valid to enable replacement
}
\examples{
# A datatable of the filtered backsplice junction read data in a sample object
#bsj.reads(sample.object) <- new.datatable

# A datatable of the filtered backsplice junction read data in each sample of an experiment
#bsj.reads(experiment.object) <- new.datatable.list

}
