% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bsjExonStats}
\alias{bsjExonStats}
\alias{bsjExonStats,circSample-method}
\alias{bsjExonStats,circExperiment-method}
\alias{plotExonStats}
\title{Calculate exon statistics}
\usage{
bsjExonStats(object, ...)

\S4method{bsjExonStats}{circSample}(object, annot, ncores = 4, ...)

\S4method{bsjExonStats}{circExperiment}(object, annot, ...)

plotExonStats(stats, size_limit = NULL, count_limit = NULL, ...)
}
\arguments{
\item{object}{A circSample or circExperiment object}

\item{...}{}

\item{annot}{The annotation object, either as an \code{EnsDb} object or a pre defined exons by transcript \code{CompressedGRangesList} object, generated from running \code{exonsBy(x = annot, by = "tx")}}

\item{ncores}{The number of threads (must be a whole number)}

\item{stats}{\code{plotExonStats}: Output from \code{bsjExonStats}}

\item{size_limit}{\code{plotExonStats}: Collapse data points where the exonic size is higher thatn the given limit. When NULL (default) data is not collapsed.}

\item{count_limit}{\code{plotExonStats}: Collapse data points where the exonic count is higher thatn the given limit. When NULL (default) data is not collapsed.}
}
\value{


Generates histograms over the exon statistics
}
\description{
\code{bsjExonStats} determines which circRNAs covers exons of the transcript models provided through the \code{annot} argument,
 and calculates overall statistics for the circRNAs in the input \code{circSample} or \code{circExperiment} object.
 The circRNA size are calculated from splicing exons, and covered exons are counted. overlapping exons in a given transcript model is merged,
 and in case of a circRNA covering multiple transcritp model, the largest size is calculated and selected.
 Be patient: It currently take minutes to complete when ncores = 5.

\code{plotExonStats} Plots crude histograms of the overall statistics calculated for exons from transcript models, covered by detected circRNAs.
}
\examples{
# Execute with circSample object using default threading (4 cores)
ahdb <- AnnotationHub()[["AH64923"]]  #  Ensembl Homo sapiens release-94
bsjExonStats(object = circ_smpl, annot = ahdb)

# Exons by transcripts can be used as annotation object
ahdb <- AnnotationHub()[["AH64923"]]
ex_tx <- exonsBy(x = ahdb, by = "tx")
bsjExonStats(object = circ_smpl, annot = ex_tx)

# Execute with circExperiment object using 10 cores
bsjExonStats(object = circ_exp, annot = ahdb, ncores = 10L)

stats <- bsjExonStats(object = circ_exp, annot = ahdb, ncores = 10L)
# Collapse circRNAs with exon sizes > 2000 and exon counts > 10, into a single column.
plotExonStats(stats, size_limit = 2000, count_limit = 10)

# Return a list of ggplot object
p <- plotExonStats(stats)

}
\author{
Kasper Thystrup Karstensen
}
