% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{bsjStats}
\alias{bsjStats}
\title{Make summary statistics of the chimeric/backspliced reads.}
\usage{
bsjStats(object, out_type = "wide", ...)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
tibble with summary statistics.
}
\description{
Make summary statistics of the chimeric/backspliced reads.
}
\examples{
bsjStats(object)

}
