% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{BSJTrack}
\alias{BSJTrack}
\title{Create track of backsplice sites}
\usage{
BSJTrack(bsj = NULL)
}
\arguments{
\item{sj}{Data on backsplice sites. Should be a GRanges object returned by readBSJout or countSumChimFile.}
}
\description{
Function to create a CustomTrack of backsplice junctions for plotting using Gviz
}
