% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{bsReadCoverage}
\alias{bsReadCoverage}
\title{Determine coverage of backspliced reads.}
\usage{
bsReadCoverage(object, bsid)
}
\arguments{
\item{object}{circSample}

\item{bsid}{Backsplice ID}
}
\description{
Determine coverage of backspliced reads.
}
