% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{c2l.ratio}
\alias{c2l.ratio}
\title{Function to calculate the circular to linear ratio for a given bsID}
\usage{
c2l.ratio(object, bsid, onlyAnnotated = TRUE, ...)
}
\arguments{
\item{object}{circSample}

\item{bsid}{Backsplice ID}

\item{onlyAnnotated}{Boolean value indicating whether to use only known linear splice junctions}
}
\description{
This function identifies linear splice sites that use donor or acceptor from a given bsID
and calculate the circ:lin ratio
}
