% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{calcCoverage}
\alias{calcCoverage}
\title{Function to calculate coverage of backspliced reads}
\usage{
calcCoverage(df, asGRanges = T)
}
\arguments{
\item{df}{Tibble of candidate backspliced reads}

\item{asGranges}{Boolean indicating whether result should be returned as GRanges obejct, otherwise a table is returned}
}
\description{
After subsetting table of candidate backsplice sites to a single backsplice sites
we can calculate the coverage.
}
