% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_data_manipulation.r
\name{categorizeBSJ}
\alias{categorizeBSJ}
\title{Function to categorize backsplice sites}
\usage{
categorizeBSJ(cbs, stranded = T)
}
\arguments{
\item{cbs}{Table (tibble) of candidate baclsplice sites}
}
\description{
Read are classified into various classes based on distance/overlap with known junctions. Need to make
function strand insensitive.
}
