% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mate.r
\name{checkChimFile}
\alias{checkChimFile}
\title{Function to test if Chimeric.out.junction file can be read without problems}
\usage{
checkChimFile(file, ...)
}
\arguments{
\item{fn}{Path to file}
}
\value{
Boolean value, any errors/warnings when reading fiurst 1000 lines.
}
\description{
Attempts to read the first 1000 lines of the file. If no errors/warnings are
encountered then it is assumed that the entire file is OK.
}
