% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mate.r
\name{checkInputFiles}
\alias{checkInputFiles}
\title{Check if amount of input files are correct, and notify if any files are missing.}
\usage{
checkInputFiles(files, single_file)
}
\arguments{
\item{files}{A character vector of one or more file names}

\item{single_file}{A logical variable, defining whether the function supports one or multiple files}

\item{dev}{Developper mode, not yet implemented, may be removed!}
}
\description{
Check if amount of input files are correct, and notify if any files are missing.
}
