% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mate.r
\name{checkPEReads}
\alias{checkPEReads}
\title{Check if both mates of a read pair are consistent with the identified backsplice junction.}
\usage{
checkPEReads(df, endTol = 5)
}
\arguments{
\item{df}{Chimeric.out.junction table returned by getCandidateBackspliceSites}

\item{endTol}{How many nucleotides on the "wrong" side of the junction can be tolerated.}
}
\value{
Logical vector
}
\description{
This function tests if the two fragments of a chimeric read fall within the
range defined defined by the backsplice junctions. Using the genome coordinates
of the acceptor and donor (columns 2 and 5 in Chimeric.out.junction), the
function checks if the region covered by the two fragments of the chimeric reads
(i.e. column 11 plus the width of the CIGAR string in column 12 and column 13
plus the width of the CIGAR string in column 14) are within the region spanned
by the acceptor and donor sites. By setting the end tollerance (endTol) you can allow for some
coverage outside the circularized region.
}
