% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_data_manipulation.r
\name{checkPEReads_old}
\alias{checkPEReads_old}
\title{Check if both mate pairs are consistent with backsplice junction.}
\usage{
checkPEReads_old(df = NULL, endTol = 5)
}
\arguments{
\item{df}{Path to a SJ.out.tab file or a data.frame.}

\item{endTol}{How many nucleotides on the "wrong" side of the junction can be tolerated.}
}
\value{
data.frame with added column of boolean values indicating whether read pair is OK.
}
\description{
Deprecated function; not particularly intuitive
}
\examples{
bsj.data <- checkPEreads(df = chi.data)
}
