% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mate.r
\name{checkVariables}
\alias{checkVariables}
\title{A error message generator, for checking classes of variables used in functions.}
\usage{
checkVariables(obj, expect_class, vari)
}
\arguments{
\item{obj}{Variable object}

\item{expect_class}{A string value of the expected class type}

\item{vari}{A string value of the function variable name}
}
\value{
Null if no error occured, otherwise returns a custom error message.
}
\description{
A error message generator, for checking classes of variables used in functions.
}
\examples{
# Will produce an error
a <- function(x){
  if (!is.null(err <- checkVariables(obj = x, req_class = "numerical", vari = "x"))) stop(err)
  return(x + 10)
}
# Will produce a costum error
a("one")
# Will execute the entire function
a(1)
}
