% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{chim.file<-}
\alias{chim.file<-}
\title{Replacement method for the chimeric file path}
\usage{
chim.file(object) <- value
}
\arguments{
\item{object}{circSample object}
}
\value{
New class object with replaced chimeric file path
}
\description{
circSample class replacement methods for replacing the chimeric file path
The replacment value must be valid to enable replacement
}
\examples{
# Replacing chimeric file path of a sample object
chim.file(sample.object) <- new.path

}
