% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{chim.file}
\alias{chim.file}
\title{Accessor for the chimeric file path}
\usage{
chim.file(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with the chimeric file path(s)
}
\description{
circSample and circExperiment class accessors for showing the chimeric file path(s) of circSample object(s)
}
\examples{
# The chimeric file path of a sample object
#chim.file(sample.object)

# The chimeric file path in each sample of an experiment
#chim.file(experiment.object)

}
