% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{cigar2covr}
\alias{cigar2covr}
\title{Function to return covered chromosomal positions from a start position and CIGAR string}
\usage{
cigar2covr(first.pos = NULL, cigar = NULL)
}
\arguments{
\item{first.pos}{Genomic position where the CIGAR string starts (from 11th or 13th column in the Chimeric.out.junction file)}

\item{cigar}{CIGAR string (from 12th or 14th column in the Chimeric.out.junction file)}
}
\description{
Function to return covered chromosomal positions from a start position and CIGAR string
}
