% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{circExperiment-class}
\alias{circExperiment-class}
\alias{circExperiment}
\title{S4 class for representing an experiment consisting of multiple samples (i.e. circSample objects).}
\value{
circExperiment-object
}
\description{
S4 class for representing an experiment consisting of multiple samples (i.e. circSample objects).
}
\section{Slots}{

\describe{
\item{\code{path}}{character Path to the directory containing output from the STAR aligner.}

\item{\code{samples}}{list A list of circSamples.}

\item{\code{name}}{character A free text field holding the name of the experiment.}
}}

\examples{
circExperiment(
    path = "path_to_data",
    name = "TEST experiment"
)

}
