\name{circular-deprecated}
\alias{circular-deprecated}
\alias{plot.edf}
\alias{lines.edf}  
\title{Deprecated Functions in Package \pkg{circular}}
\description{
  These functions are provided for compatibility with older versions of
  the package only, and may be defunct as soon as the next release.
}

\details{
Functions deprecated are

\code{plot.edf}

\code{lines.edf}    

}  

  
\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
