% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{constructBsId}
\alias{constructBsId}
\title{Construct unique backsplice junction IDs}
\usage{
constructBsId(object, cores = 1L)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{cores}{Only supported on unix! Number of cores used for parallel processing for circExperiment objets.}
}
\value{
A new circSample or circExperiment object, with a bsID column in the bsj.read datatable.
}
\description{
Generate unique readID from information of the candidate backsplice junction read data.
}
\examples{
constructBsId(object)

}
