% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{constructSJDB}
\alias{constructSJDB}
\title{Construct a database of known splice junctions from a database of annotated genes.}
\usage{
constructSJDB(annotationDB, force = FALSE)
}
\arguments{
\item{annotationDB}{A TxDb or EnsDb object}

\item{force}{Boolean (default = FALSE) indicating whether or not to overwrite any cached version splice junction database.}
}
\description{
This function takes an annotation database, either downloaded from AnnotationHub
or generated using makeTxDbFromBiomart (GenomicFeatures) and extracts all junction sites and transcript termini.
}
