\name{conversion.circular}
\alias{conversion.circular}
\title{Unit of Measure Conversion for Circular Data}
\description{
Unit of Measure Conversion for Circular Data
}
\usage{
conversion.circular(x, units = c("radians", "degrees"))
}

\arguments{
  \item{x}{an object of class \code{\link{circular}}.}
  \item{units}{unit of the transformed data.}
}

}
\value{
  an object of class \code{\link{circular}} with the specified unit of
  measure.
}

\author{Claudio Agostinelli}

\seealso{
\code{\link{deg}} and \code{\link{rad}}
}

\keyword{misc}
