\name{Coope}
\alias{x.coope}
\alias{y.coope}

\title{Coope dataset  }

\usage{
data(coope)
}

\description{
A dataset taken from the paper of Coope (1993). 
}

\format{
\code{x.coope} and \code{y.coope} are vectors of length  8.
}

\source{
Coope, I. (1993). Circle fitting by linear and non-linear least squares. 
Journal of Optimization Theory and Applications, 76, 381-388. 
}

\keyword{datasets}
