% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{count.file<-}
\alias{count.file<-}
\title{Replacement method for the linear read count data file path}
\usage{
count.file(object) <- value
}
\arguments{
\item{object}{circSample object}
}
\value{
New class object with replaced linear read count data file path
}
\description{
circSample class replacement methods for replacing the linear read count data file path
The replacment value must be valid to enable replacement
}
\examples{
# Replacing linear read count data file path of a sample object
count.file(sample.object) <- new.path

}
