% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{count.file}
\alias{count.file}
\title{Accessor for the linear read count data file path}
\usage{
count.file(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with the linear read count data file path(s)
}
\description{
circSample and circExperiment class accessors for showing the linear read count data file path(s) of circSample object(s)
}
\examples{
# The linear read count data file path of a sample object
#count.file(sample.object)

# The Linear splice junctions file path in each sample of an experiment
#count.file(experiment.object)

}
