% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{countSumChimFile}
\alias{countSumChimFile}
\title{Summarize the read count for each backsplice junction}
\usage{
countSumChimFile(df, asGRanges = F)
}
\arguments{
\item{df}{Table (tibble) of chimeric reads}

\item{asGRanges}{Boolean indication if results should be returned as a GRanges}
}
\description{
Summarize the read count for each backsplice junction
}
