\name{deg}    
\title{Degrees}
\alias{deg}

\description{
Converts radians to degrees. 
}

\usage{
deg(x)
}

\arguments{
\item{x}{vector or matrix of radian measurements.}
}

\value{
Returns a vector or matrix of degree measurements corresponding to the
data in radians.
}

\details{This function is available for compatibility with the CircStats
  package; please use \code{\link{conversion.circular}}.}

\author{Claudio Agostinelli and Ulric Lund}

\seealso{
  \code{\link{conversion.circular}} and \code{\link{rad}}
}

\keyword{math}
