% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{determineAmbiguousHits}
\alias{determineAmbiguousHits}
\title{Analyses the provided donors and acceptors to determine if there are any junctions that cannot be unequivocally determined by the nearest function}
\usage{
determineAmbiguousHits(acceptors, donors)
}
\description{
Analyses the provided donors and acceptors to determine if there are any junctions that cannot be unequivocally determined by the nearest function
}
