% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{exp.mat}
\alias{exp.mat}
\title{Accessor for expression data from circExperiment}
\usage{
exp.mat(object, format = "long", normFactors = NULL, ...)
}
\arguments{
\item{object}{circExperiment object}

\item{format}{Should result be returned as a "long" (fast) or "wide" (slow) table}
}
\value{
A tibble with expression values
}
\description{
circExperiment class accessors for expression data from a circExperiment
}
\examples{
# The samples of an experiment
#exprs(circExperiment)

}
