% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{findNearest}
\alias{findNearest}
\title{Candidate backsplice data are split into donors and acceptors as an GRanges object,
the annotation data are screened, to determine which individual donor and acceptor site maps to a nearby known junction.
The candidate backsplice data that overlaps or are near a known junction site are annotated and returned.
Following annotation are used for donors and acceptors:
* queryHit: The dataset postioin of a given backsplice candidate read.
* jID: The closest known junction unqiue ID (provided during 'constructSJDB')
* type: The closest known junction type (acceptor or donor)
* strand: The closest known junction strand position
* ToNearestJ: Minimum distance of the read donor or acceptor position to the nearest known donor or acceptor site.}
\usage{
findNearest(juncCand, kj, type, cbs)
}
\description{
Candidate backsplice data are split into donors and acceptors as an GRanges object,
the annotation data are screened, to determine which individual donor and acceptor site maps to a nearby known junction.
The candidate backsplice data that overlaps or are near a known junction site are annotated and returned.
Following annotation are used for donors and acceptors:
* queryHit: The dataset postioin of a given backsplice candidate read.
* jID: The closest known junction unqiue ID (provided during 'constructSJDB')
* type: The closest known junction type (acceptor or donor)
* strand: The closest known junction strand position
* ToNearestJ: Minimum distance of the read donor or acceptor position to the nearest known donor or acceptor site.
}
