% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{firstread.firststrand<-}
\alias{firstread.firststrand<-}
\title{Replacement method for the strandedness protocol of sample library preparation}
\usage{
firstread.firststrand(object) <- value
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
New class object with replaced logical firstread-firststrand vector
}
\description{
circSample and circExperiment class replacement methods for replacing the boolean value stating if the library preparationion protocol yielded firstread-firststrand reads
The replacment value must be valid to enable replacement
}
\examples{
# Replacing boolean value of whether reads are firstread-firststrand in a sample object
#firstread.firststrand(sample.object) <- new.logical

# Replacing boolean value whether reads are firstread-firststrand in each sample of an experiment
#firstread.firststrand(experiment.object) <- new.logical.vector

}
