% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{firstread.firststrand}
\alias{firstread.firststrand}
\title{Accessor for the strandedness protocol of sample library preparation}
\usage{
firstread.firststrand(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Logical vector with a boolean value stating if the library preparationion protocol yielded firstread-firststrand reads
}
\description{
circSample and circExperiment class accessors for showing whether the sample are prepared using e.g. dUTP- or unstranded protocol
}
\examples{
# The logical value whether the protocol yielded firstread-firststrand reads in a sample object
firstread.firststrand(sample.object)

# The logical value whether the protocol yielded firstread-firststrand reads in each sample of an experiment
firstread.firststrand(experiment.object)

}
