\name{fisherB2}
\alias{fisherB2}
\alias{fisherB2c}

\title{B.2 Measurements of long-axis orientation of 133 feldspar laths in basalt}

\usage{
data(fisherB2)
data(fisherB2c)
}

\description{
Measurements of long-axis orientation of 133 feldspar laths in basalt
}
\format{
\code{fisherB2} is a vector of 133 observations (in degrees).
\code{fisherB2c} contains the same observations in a circular objects.
}

\source{
Smith, N.M. (1988) Reconstruction of the Tertiary drainage systems of the Inverell region. Unpublished B.Sc. (Hons.) thesis, Department of Geography, University of Sydney, Australia.

This dataset (set 28-6-1co.prn) was kindly supplied by Ms Nicola Smith to Prof. N.I. Fisher.
}
\seealso{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press. Pag. 240.
}

\examples{
data(fisherB2c)
plot(fisherB2c)
}

\keyword{datasets}
