\name{fisherB3}
\alias{fisherB3}
\alias{fisherB3c}

\title{B.3 Movements of turtles}

\usage{
data(fisherB3)
data(fisherB3c)
}

\description{
Measurements of the directions taken by 76 turtles after treatment.
}
\format{
\code{fisherB3} is a vector of 76 observations (in degrees).
\code{fisherB3c} contains the same observations in a circular objects.
}

\source{
Stephens, M.A. (1969) Techniques for directional data. Thechnical Report #150, Department of Statistics, Stanford University, Stanford, CA.
}
\seealso{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press. Pag. 241.
}

\examples{
data(fisherB3c)
plot(fisherB3c)
}

\keyword{datasets}
