\name{fisherB4}
\alias{fisherB4}
\alias{fisherB4c}

\title{B.4 Directional preferences of starhead topminnows}

\usage{
data(fisherB4)
data(fisherB4c)
}

\description{
Sun compass orientations of 50 starhead topminnows, measured under heavily overcast conditions.
}
\format{
\code{fisherB4} is a vector of 50 observations (in degrees).
\code{fisherB4c} contains the same observations in a circular objects.
}

\source{
Goodyear (1970) Terrestrial and aquatic orientation in the Starhead Topminnow, Fundulus notti. Science 168, 603-5. Figure 1D.
}
\seealso{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press. Pag. 241.
}

\examples{
data(fisherB3c)
plot(fisherB3c)
}

\keyword{datasets}
