\name{fisherB5}
\alias{fisherB5}
\alias{fisherB5c}

\title{B.5 Measurements of long-axis orientation of 164 feldspar laths in basalt}

\usage{
data(fisherB5)
data(fisherB5c)
}

\description{
Measurements of long-axis orientation of 164 feldspar laths in basalt
}
\format{
\code{fisherB5} is a vector of 164 observations (in degrees).
\code{fisherB5c} contains the same observations in a circular objects.
}

\source{
Smith, N.M. (1988) Reconstruction of the Tertiary drainage systems of the Inverell region. Unpublished B.Sc. (Hons.) thesis, Department of Geography, University of Sydney, Australia.

This dataset (set 24-6-5co.prn) was kindly supplied by Ms Nicola Smith to Prof. N.I. Fisher.
}
\seealso{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press. Pag. 242.
}

\examples{
data(fisherB5c)
plot(fisherB5c)
}

\keyword{datasets}
