% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{gb<-}
\alias{gb<-}
\title{Replacement method for genome build}
\usage{
gb(object) <- value
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
New class object with replaced genome build(s)
}
\description{
circSample and circExperiment class replacement methods for replacing the genome build(s) of circSample object(s)
The replacment value must be valid to enable replacement
}
\examples{
# Replacing genome build of a sample object
#gb(sample.object) <- new.gb

# Replacing genome build in each sample of an experiment
#gb(experiment.object) <- new.gb.vector

}
