% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{generateJunctionMotifs}
\alias{generateJunctionMotifs}
\title{Generate junction motifs}
\usage{
generateJunctionMotifs(object, genome_seq, cores = 1L)
}
\arguments{
\item{object}{A circSample or circExperiment object}

\item{genome_seq}{A BSGenome object used for determining the sequences for splice motifs}
}
\value{
an updated circSample or circExperiment object
}
\description{
Generates splice motif sequences and adds them t the backsplice junction datatable
}
\examples{
# Do not run
# Setup circSample or circExperiment object
#object <- circSample(...)
# Load relevant data
#...
# Generate motifs
#object <- generateJunctionMotifs(object)
}
