% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.r
\name{getCandidateBackspliceSites}
\alias{getCandidateBackspliceSites}
\title{Read candidate backsplice sites from Chimeric.out.junction file}
\usage{
getCandidateBackspliceSites(
  fn,
  backSpliceDist = 1e+05,
  seqAcrossSj = TRUE,
  libType = "fr-firststrand",
  verbose = FALSE
)
}
\arguments{
\item{fn}{Path to a Chimeric.out.junction file.}

\item{backSpliceDist}{The maximum allowed distance between donor and acceptor sites.}

\item{seqAcrossSj}{Boolean indicating whether to return all backspliced reads or just reads where one of the mates span the candidate backsplice site.}

\item{libType}{Character string indicating which stranded library protocol was used to generate the data, must be 'fr-firststrand' for e.g. dUTP method or 'fr-secondstrand'.}

\item{verbose}{Inlcude basic statstics on the chimeric readsm ust be logical (default FALSE).}
}
\value{
tibble of reads covering backsplice sites
}
\description{
\code{getCandidateBackspliceSites} reads entire Chimeric.out.junction file and returns reads consistent with backsplicing.
Candidate backsplice junctions are identified by applying the following filters
\enumerate{
  \item Donor and acceptor are on the same chromosome
  \item Donor and acceptor are on the same strand
  \item The acceptor is upstream of donor (On plus strand: acceptor position < donor position ; On minus strand: donor position < acceptor position)
  \item Distance between donor and acceptor is no more than 100.000 bases (by default, can be changed)
}
}
