% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{getKnownJunctions}
\alias{getKnownJunctions}
\title{Generate database of known splice junction sites from a TxDb- or EnsDb-database}
\usage{
getKnownJunctions(annotationDB)
}
\arguments{
\item{annotationDB}{TxDb or EnsDb object}
}
\description{
Generate database of known splice junction sites from a TxDb- or EnsDb-database
}
