% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{guessPE}
\alias{guessPE}
\title{Guess whether data are paired ended or not by parsing the BAM file header}
\usage{
guessPE(bf)
}
\arguments{
\item{bf}{path to bam file}
}
\value{
Logical vector of length 1, indicating if sample is paired ended (TRUE), single ended (FALSE), or not able to determine (NULL)
}
\description{
Guess whether data are paired ended or not by parsing the BAM file header
}
\examples{
guessPE(path.to.bam.file)
}
