\name{I.0}
\title{Zeroth Order Bessel Function of the First Kind}
\alias{I.0}

\description{
An alias of \code{besselI(x, nu=0)}.
}

\usage{
I.0(x)
}

\arguments{
\item{x}{non-negative numerical value at which to evaluate the Bessel function.}
}

\value{
Returns the zeroth order Bessel function of the first kind evaluated at a specified real number.
}

\seealso{
\code{\link{besselI}}.
}

\keyword{math}

