\name{I.1}
\title{First Order Bessel Function of the First Kind}
\alias{I.1}

\description{
An alias of \code{besselI(x, nu=1)}.
}

\usage{
I.1(x)
}

\arguments{
\item{x}{non-negative numerical value at which to evaluate the Bessel function.}
}

\value{
Returns the first order Bessel function of the first kind, evaluated at a specified real number.
}


\seealso{
\code{\link[base]{besselI}}. 
}

\keyword{math}

