% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{indexJunctions}
\alias{indexJunctions}
\title{FURTHER FUNCTIONALITY should be implemented if we wish to determine lariats and intronic reads -> further functionality are currently commented out!!
Determines which known backsplice donor and acceptor junction candidates can be rescued and indexes these.}
\usage{
indexJunctions(donorSubset, acceptorSubset, type)
}
\description{
FURTHER FUNCTIONALITY should be implemented if we wish to determine lariats and intronic reads -> further functionality are currently commented out!!
Determines which known backsplice donor and acceptor junction candidates can be rescued and indexes these.
}
