% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{inferExons}
\alias{inferExons}
\title{Infer the exons structure of a circular RNA}
\usage{
inferExons(bsid, eBt)
}
\arguments{
\item{bsid}{A valid backsplice ID}

\item{eBt}{A GRangesList of exons-by-transcript, generated using the exonsBy-function.}
}
\value{
GRanges object or list of GRanges obejcts giving the inferred exon structure of a circRNA identified by the supplied backsplice ID.
}
\description{
The genomic region covered by a backsplice junction is used to identify overlapping transcript models
and subsequently, whether the nucleotides adjacent to the donor and acceptor positions overlap with known exon boundaries.
If both positions overlap with known exon termini, the resulting exon structure is returned. Otherwise, NULL is returned.
}
