% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{is.empty}
\alias{is.empty}
\title{Check for empty values}
\usage{
is.empty(object)
}
\arguments{
\item{object}{The object that are to be checked.}
}
\value{
A logical value that states whether the object is considered empty.
}
\description{
Check if the object has an empty value, does not include checking empty strings.
}
\examples{
# Test that turns out TRUE
a <- NULL
b <- character()
is.empty(a)
is.empty(b)
is.empty(list())  # is TRUE as length = 0
is.empty(c())

# Test that turns out FALSE
is.empty("")
is.empy(NA)
}
