% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{junctionMotif}
\alias{junctionMotif}
\title{Extract junction motif(s) from genome}
\usage{
junctionMotif(bsids, g = Hsapiens)
}
\arguments{
\item{bsids}{An character vector with bsIDs.}

\item{g}{A BS genome object. Default Hsapiens.}
}
\value{
Character vector with the junction motifs.
}
\description{
This function takes a vector of bsIDs and a BSgenome-object and constructs junction-motifs of the backsplice junctions.
}
