% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{junctionsToGRanges}
\alias{junctionsToGRanges}
\title{Simple wrapper for generating a GRanges object, it supports the same input as 'GenomicRanges::Ranges', i.e. addition of metadata columns.
As a minimum a chromosome name and aligned strand must be provided, in addition two of either must be provided of following; a 'start', 'end', or 'width'}
\usage{
junctionsToGRanges(
  seqnames,
  start = NULL,
  end = NULL,
  width = NULL,
  strand,
  ...
)
}
\description{
Simple wrapper for generating a GRanges object, it supports the same input as 'GenomicRanges::Ranges', i.e. addition of metadata columns.
As a minimum a chromosome name and aligned strand must be provided, in addition two of either must be provided of following; a 'start', 'end', or 'width'
}
