% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{knownJunctionSites}
\alias{knownJunctionSites}
\title{Determine whether reads in a bsj.read-table uses known splice junctions}
\usage{
knownJunctionSites(df)
}
\arguments{
\item{df}{BSJ data table}
}
\value{
A vector that is as long as there are rows in the input.
}
\description{
Determine whether reads in a bsj.read-table uses known splice junctions
}
\examples{
...
}
