% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{label}
\alias{label}
\title{Accessor for sample label}
\usage{
label(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with sample label(s)
}
\description{
circSample and circExperiment class accessors for showing the sample label(s)
}
\examples{
# The label of a sample object
label(sample.object)

# The label in each sample of an experiment
label(experiment.object)

}
