% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{locateSamples}
\alias{locateSamples}
\title{Method for automatically detecting and setting up all samples in an experiment.}
\usage{
locateSamples(
  object,
  organism = "Homo sapiens",
  gb = "hg38",
  firstread.firststrand = T,
  paired.end = T
)
}
\value{
circExperiment object with populated sample-slot.
}
\description{
Method for automatically detecting and setting up all samples in an experiment.
}
\examples{
locateSamples(experiment.object)

}
