% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{log.file<-}
\alias{log.file<-}
\title{Replacement method for the STAR alignment Log file path}
\usage{
log.file(object) <- value
}
\arguments{
\item{object}{circSample bject}
}
\value{
New class object with replaced STAR alignment Log file path
}
\description{
circSample class replacement methods for replacing the STAR alignment Log file path
The replacment value must be valid to enable replacement
}
\examples{
# Replacing STAR alignment Log file path of a sample object
log.file(sample.object) <- new.path

}
