% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{log.file}
\alias{log.file}
\title{Accessor for the STAR alignment Log file path}
\usage{
log.file(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with the STAR alignment Log file path(s)
}
\description{
circSample and circExperiment class accessors for showing the STAR alignment Log file path(s) of circSample object(s)
}
\examples{
# The STAR alignment Log file path of a sample object
log.file(sample.object)

# The STAR alignment Log file path in each sample of an experiment
log.file(experiment.object)

}
