% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{lsj.counts<-}
\alias{lsj.counts<-}
\title{Replacement method for summarized data on linear splice sites}
\usage{
lsj.counts(object) <- value
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
New class object with replaced linear splice junction count data
}
\description{
circSample and circExperiment class replacement methods for replacing the linear splice junction count data
The replacment value must be valid to enable replacement
}
\examples{
# A datatable of the linear splice junction count data in a sample object
#lsj.counts(sample.object) <- new.datatable

# A datatable of the linear splice junction count data in each sample of an experiment
#lsj.counts(experiment.object) <- new.datatable.list

}
