% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{lsj.counts}
\alias{lsj.counts}
\title{Accessor for summarized data on linear splice sites}
\usage{
lsj.counts(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Datatable(s) of linear splice junction count data
}
\description{
circSample and circExperiment class accessors for showing the linear splice junction count data
}
\examples{
# A datatable of the linear splice junction count data in a sample object
lsj.counts(sample.object)

# A datatable of the linear splice junction count data in each sample of an experiment
lsj.counts(experiment.object)

}
