% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{name<-}
\alias{name<-}
\title{Replacement method for experiment name}
\usage{
name(object) <- value
}
\arguments{
\item{object}{circExperiment object}
}
\value{
New class object with replaced experiment name
}
\description{
circExperiment class replacement methods for replacing the experiment name
The replacment value must be valid to enable replacement
}
\examples{
# Replacing the experiment name
name(experiment.object) <- new.name

}
