% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{name}
\alias{name}
\title{Accessor for experiment name
circExperiment class accessors for showing the experiment name}
\usage{
name(object)
}
\arguments{
\item{object}{circExperiment object}
}
\value{
Character string with the experiment name
}
\description{
Accessor for experiment name
circExperiment class accessors for showing the experiment name
}
\examples{
# The experiment name
#name(experiment.object)

}
