\name{ncfrog}
\alias{ncfrog}
\alias{ncfrog.rad}

\title{Northern Cricket Frog}

\usage{
data(ncfrog)
}

\description{
 In an experiment due to Ferguson et al. (1967) a number of northern
 cricket frogs (Acris crepitans) were collected from the mud flats of an
 abandoned stream meandering near Indianola, Mississippi, and taken to a test pen
lying to the north west of the collection point. After 30 hours of
enclosure within a dark environmental chamber, 14 of them were released and
the directions taken by these frogs recorded. 0 degrees means North.
}
\format{
\code{ncfrog} is a vector of 14 observations (in degrees).
\code{ncfrog.rad} contains the same observations in radians (pi/180).
}

\source{
 Collett, D. (1980)
 Outliers in Circular Data
 \emph{Applied Statistics}
 \bold{29}, 1, 50--57.
}

\seealso{
  Ferguson, D.E, Landreth, H.F. and McKeown, J.P. (1967)
  Sun compass orientation of the northern cricket frog, \emph{Acris crepitans}.
  \emph{Anim. Behav.},
  \bold{14}, 45--53.
}


%\examples{}

\keyword{datasets}
