% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{organism}
\alias{organism}
\title{Accessor for organism}
\usage{
organism(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with organism(s)
}
\description{
circSample and circExperiment class accessors for showing the organism(s) of cicSample object(s)
}
\examples{
# The organism of a sample object
organism(sample.object)

# The organism in each sample of an experiment
organism(experiment.object)
}
