% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{paired.end<-}
\alias{paired.end<-}
\title{Replacement method for the library preparation protocol realtive to end-pairing}
\usage{
paired.end(object) <- value
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
New class object with replaced logical paired-end vector
}
\description{
circSample and circExperiment class replacement methods for replacing the boolean value stating if the library preparationion protocol yielded paired-end reads
The replacment value must be valid to enable replacement
}
\examples{
# Replacing boolean value whether reads are paired-end reads in a sample object
paried.end(sample.object) <- new.logical

# Replacing logical value whether reads are paired-end reads in each sample of an experiment
paried.end(experiment.object) <- new.logical.vector

}
