% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{paired.end}
\alias{paired.end}
\title{Accessor for the library preparation protocol realtive to end-pairing}
\usage{
paired.end(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Logcal vector with a boolean value stating if the library preparationion protocol yielded paired-end reads
}
\description{
circSample and circExperiment class accessors for showing whether the sample are prepared as paired-end or single-end
}
\examples{
# The logical value whether the protocol yielded paired-end reads in a sample object
paried.end(sample.object)

# The logical value whether the protocol yielded paired-end reads in each sample of an experiment
paried.end(experiment.object)

}
