% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{parseCIGAR}
\alias{parseCIGAR}
\title{#' Construct bsID from genomic coordinates
#'
#' @param gb Genome build
#' @export
constructBsId <- function(df, gb){
  bsID <- paste(gb, df$X1, df$X2, df$X5, df$X3, sep = ":")
  return(bsID)
}
Calculate distance covered by the alignment from a CIGAR string}
\usage{
parseCIGAR(cig, returnLength = T)
}
\arguments{
\item{cig}{The CIGAR string.}
}
\value{
value indicating number of genomic nucleotides covered by the alignment.
}
\description{
### Do we wish to export?? (only if used outside of "wrapper functions")
}
\examples{
bla bla
parseCIGAR(cig = "77M107N24M-135p4M107N77M20S")
}
