% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{parseDataDir}
\alias{parseDataDir}
\title{Parse a directory to identify samples for processing in the circulaR pipeline}
\usage{
parseDataDir(dir, PE = NULL, frfs = NULL, org = NULL, geno = NULL)
}
\arguments{
\item{dir}{The path to directory where STAR output files are stored}
}
\value{
Unknown
}
\description{
Parse a directory to identify samples for processing in the circulaR pipeline
}
\examples{
...
}
