% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{path}
\alias{path}
\title{Accessor for main directory path
circExperiment class accessors for showing the experiment main path}
\usage{
path(object, ...)
}
\arguments{
\item{object}{circExperiment object}
}
\value{
Character string with the main directory path
}
\description{
Accessor for main directory path
circExperiment class accessors for showing the experiment main path
}
\examples{
# The main path to the experiment
path(experiment.object)

}
