\name{pigeons}
\docType{data}
\alias{pigeons}
\title{Initial orientation of displaced homing pigeons}
\description{This data set has 108 rows and 2 columns. The observations
  are the vanishing bearings of homing pigeons displaced and released at
  two unfamiliar locations. The data are pooled with respect to the home
  direction (home direction set in 360 grades). 
}
\usage{data(pigeons)}
\format{This data frame contains the following columns:
  \describe{
    \item{\code{treatment}}{, a factor with levels: \emph{c}, control pigeon
      (unmanipulated); \emph{v1}, pigeons subjected to bilateral section
      of the ophthalmic branch of the trigeminal nerve; \emph{on},
      pigeons subjected to bilateral section of the olfactory nerve}
    \item{\code{bearing}}{, vanishing bearing of each bird in degrees}
  }
}
\references{
       Gagliardo A., Ioale' P., Savini M., and Wild M. (2008). Navigational
       abilities of homing pigeons deprived of olfactory or trigeminally
       mediated magnetic information when young. \emph{J. Exp. Biol.},
       \bold{211}:2046--2051.
}
\keyword{datasets}
