% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plotAll}
\alias{plotAll}
\title{Make homemade sashimi plot}
\usage{
plotAll(
  DB = txdb,
  sj = sj.data,
  bsj = bsj.data,
  region = region.of.interest,
  gid = g,
  ...
)
}
\arguments{
\item{DB}{TxDb of transcripts}

\item{sj}{Linear splice junction data (output from readSJ)}

\item{bsj}{Backsplice junction data (output from readBSJ)}

\item{region}{GRanges object describing the region of interest}

\item{gid}{Ensembl gene id to plot grid.newpage}
}
\description{
Plot transcript models and show linear and backsplice jjunctions.
}
