% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plotBSJ}
\alias{plotBSJ}
\title{Plot backsplice junctions stored in a GRanges object.}
\usage{
plotBSJ(df = NULL, rangex, line.scaling = 0.5)
}
\arguments{
\item{df}{yada yada yada}

\item{rangex}{see plotTranscripts}

\item{line.scaling}{Factor for overall scaling of line widths.}
}
\description{
Plot backsplice junctions stored in a GRanges object.
}
