% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plotBSJrepeats}
\alias{plotBSJrepeats}
\title{Zoomed in view of the donor and acceptor sites}
\usage{
plotBSJrepeats(
  df = candidate.bsj,
  rid = NULL,
  size = 10,
  l.fontsize = 9,
  unstranded = F
)
}
\arguments{
\item{df}{Data.frame of candidate backspliced reads}

\item{rid}{Name of read to focus on}

\item{size}{Number of nucleotides adjacent to the junctions}

\item{l.fontsize}{Sice of the font used for legend}
}
\description{
Zoomed in view of the donor and acceptor sites
}
