% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plotCoverage}
\alias{plotCoverage}
\title{Function to plot coverage as barplot.}
\usage{
plotCoverage(df = NULL, rangex)
}
\arguments{
\item{df}{GRanges object returned by calcCoverage}

\item{rangex}{The range in genomic coordinates for the region of interest. This is important for correct alignment of the tx model plot with splice junction plot.}
}
\description{
Function to plot coverage as barplot.
}
