% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plotSJ}
\alias{plotSJ}
\title{Plot linear splice sites stored in a GRanges object.}
\usage{
plotSJ(df = NULL, rangex, line.scaling = 0.5, trimJ = T)
}
\arguments{
\item{df}{yada yada yada}

\item{rangex}{see plotTranscripts}

\item{line.scaling}{Factor for overall scaling of line widths.}

\item{trimJ}{Boolean, should junctions with donor/acceptor outside the range defined in 'rangex' be excluded.}
}
\description{
Plot linear splice sites stored in a GRanges object.
}
