% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plotTxModels}
\alias{plotTxModels}
\title{Plot transcript models.}
\usage{
plotTxModels(db = NULL, g = NULL, tid = NULL, showDirection = T, rangex = NULL)
}
\arguments{
\item{db}{A db containing transcript models from AnnotationHub/Ensembl.}

\item{g}{Specify which gene to plot (Ensembl Gene ID).}

\item{showDirection}{Boolean indicating wheter to add chevrons to intronic regions to indicate direction of the transcript.}

\item{rangex}{The range in genomic coordinates for the region of interest. This is important for correct alignment of the tx model plot with splice junction plot.}
}
\description{
Using data stored in an EnsDB from AnnotationHub.
}
