% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{pruneChromosomes}
\alias{pruneChromosomes}
\title{Prune chromosomes}
\usage{
pruneChromosomes(db, chromosomes = NULL)
}
\arguments{
\item{db}{An object containing a Seqinfo class, for chromosomal values.}

\item{chromosomes}{A vector of chromosomes to keep, must be `integer`, `character` or `NULL` (default). Alterantive provide the value "standard" No chromosomes are removed if set to `NULL`.}
}
\value{
Same as input, but with certin chromosomes removed.
}
\description{
Define which chromosomes to use from genomic annotation object, default `NULL` does not exclude any chromosomes
}
