\name{rad}    
\title{Radians}
\alias{rad}

\description{
Converts degrees to radians. 
}

\usage{
rad(x)
}

\arguments{
\item{x}{vector or matrix of degree measurements.}
}

\value{
Returns a vector or matrix of radian measurements corresponding to the data in degrees.
}

\details{This function is available for compatibility with the CircStats
  package, please use \code{\link{conversion.circular}}.}

\author{Claudio Agostinelli and Ulric Lund}

\seealso{
\code{\link{conversion.circular}} and \code{\link{deg}}
}

\keyword{math}
