% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{readBSJdata}
\alias{readBSJdata}
\title{Read the chimeric data of a circSample or circExperiment.}
\usage{
readBSJdata(
  object,
  chromosomes = c(1:22, "X", "Y"),
  cores = 1L,
  maxGenomicDist = 1e+05,
  onlySpanning = TRUE,
  removeBadPairs = T,
  ...
)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{chromosomes}{A vector of (valid) chromosome names that is used to restrict the chimeric read data}

\item{cores}{Only supported on unix! Number of cores used for parallel processing for circExperiment objets.}

\item{maxGenomicDist}{Maximal allowed size between donor and acceptor (defaults to 100.000 nt).}

\item{onlySpanning}{Return only BSJ spanning or all (spanning and encompassing) reads.}
}
\value{
Same as input object, but populated slot for bsj data.
}
\description{
The function takes a circSample or circExperiment object and reads the chimeric data for samples described in these objects.
}
\examples{
readBSJdata(testExperiment)
}
