% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{readChimFile}
\alias{readChimFile}
\title{Wrapper for easy reading of Chimeric.out.junction file}
\usage{
readChimFile(file, n_max = Inf, ...)
}
\arguments{
\item{file}{Path to file.}

\item{...}{
  Arguments passed on to \code{\link[readr:read_delim]{readr::read_tsv}}
  \describe{
    \item{\code{n_max}}{Maximum number of records to read.}
  }}
}
\value{
Tibble data frame with data on chimeric reads
}
\description{
This function uses read_tsv from the tidyverse to read the 14-coulmn file.
}
\examples{
#tmp <- readChimFile(fn = "Chimeric.out.junction")
}
