% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{readLSJdata}
\alias{readLSJdata}
\title{Read the data on reads spanning linear splice junctions}
\usage{
readLSJdata(object, chromosomes = c(1:22, "X", "Y"), cores = 1L, ...)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{chromosomes}{A vector of (valid) chromosome names that is used to restrict the chimeric read data}

\item{cores}{Only supported on unix! Number of cores used for parallel processing for circExperiment objets.}
}
\value{
Same as input object, but populated slot for lsj counts.
}
\description{
The function takes a circSample or circExperiment object and reads the chimeric data for samples described in these objects (SJ.out.tab file).
}
\examples{
readLSJdata(object)

}
