% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{rescueAmbiguous}
\alias{rescueAmbiguous}
\title{The function used to identify nearest known junction ('IRange::nearest') reports the gap between two genomic positions and not the distance.
This means that a position (x) can be reported as nearest to three positions (x-1, x, x+1). If such ambiguities arise by using the nearest function,
we need to determine if the results can be disambiguated. This function will return reads where:
closest.donor.type == donor & closest.acceptor.type == acceptor
AND
shiftAcceptorToNearestJ == 0 | shiftDonorToNearestJ == 0
AND
donor/acceptor pairs that are closest to known junctions}
\usage{
rescueAmbiguous(acceptors, donors, ambiguousHits)
}
\description{
The function used to identify nearest known junction ('IRange::nearest') reports the gap between two genomic positions and not the distance.
This means that a position (x) can be reported as nearest to three positions (x-1, x, x+1). If such ambiguities arise by using the nearest function,
we need to determine if the results can be disambiguated. This function will return reads where:
closest.donor.type == donor & closest.acceptor.type == acceptor
AND
shiftAcceptorToNearestJ == 0 | shiftDonorToNearestJ == 0
AND
donor/acceptor pairs that are closest to known junctions
}
