% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{resolveReads}
\alias{resolveReads}
\title{Compares the full backsplice candidate set with reads that are ambiguously mapped to the known junctions,
and deselcts all ambiguous reads. The ambiguous reads that can be rescued are then defined and included.
Finally the selected backsplice sites are fully annotated and returned.}
\usage{
resolveReads(cbs, ambiguousHits, toRescue, donors, acceptors)
}
\description{
Compares the full backsplice candidate set with reads that are ambiguously mapped to the known junctions,
and deselcts all ambiguous reads. The ambiguous reads that can be rescued are then defined and included.
Finally the selected backsplice sites are fully annotated and returned.
}
