% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{sample.id<-}
\alias{sample.id<-}
\title{Replacement method for sample.id}
\usage{
sample.id(object) <- value
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
New class object with replaced sample.id(s)
}
\description{
circSample and circExperiment class replacement methods for replacing the sample.id(s) of cicSample object(s)
The replacment value must be valid to enable replacement
}
\examples{
# Replacing sample.id of a sample object
#sample.id(sample.object) <- sample.id

# Replacing sample.id in each sample of an experiment
#sample.id(experiment.object) <- sample.id.vector

}
