% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{sample.id}
\alias{sample.id}
\title{Accessor for sample.id}
\usage{
sample.id(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with sample.id(s)
}
\description{
circSample and circExperiment class accessors for showing the sample.id(s) of cicSample object(s)
}
\examples{
# The sample.id of a sample object
sample.id(sample.object)

# The sample.id in each sample of an experiment
sample.id(experiment.object)

}
