% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\name{samples,circExperiment-method}
\alias{samples,circExperiment-method}
\title{Accessor for experiment samples}
\usage{
\S4method{samples}{circExperiment}(object)
}
\arguments{
\item{object}{circExperiment object}
}
\value{
List of samples stored in the circExperiment object
}
\description{
circExperiment class accessors for showing the circSample objects
}
\examples{
# The samples of an experiment
samples(experiment.object)

}
