% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{samples<-}
\alias{samples<-}
\title{Replacement method for samples}
\usage{
samples(object) <- value
}
\arguments{
\item{object}{circExperiment object}
}
\value{
New class object with replaced circSample's
}
\description{
circExperiment class replacement methods for replacing circSample's
The replacment value must be valid to enable replacement
}
\examples{
# Replacing samples in an experiment
#samples(experiment.object) <- new.circSample.list

}
