% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_data_manipulation.r
\name{shiftAlignment}
\alias{shiftAlignment}
\title{Function to shift alignment to overlap with known junctions}
\usage{
shiftAlignment(cbs)
}
\arguments{
\item{cbs}{Table (tibble) of candidate backspliced reads}
}
\description{
Function to shift alignment to overlap with known junctions
}
