% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{sj.file<-}
\alias{sj.file<-}
\title{Replacement method for the Linear splice junctions file path}
\usage{
sj.file(object) <- value
}
\arguments{
\item{object}{circSample object}
}
\value{
New class object with replaced Linear splice junctions file path
}
\description{
circSample class replacement methods for replacing the Linear splice junctions file path
The replacment value must be valid to enable replacement
}
\examples{
# Replacing Linear splice junctions file path of a sample object
#sj.file(sample.object) <- new.path

}
