% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{sj.file}
\alias{sj.file}
\title{Accessor for the Linear splice junctions file path}
\usage{
sj.file(object)
}
\arguments{
\item{object}{circSample or circExperiment object}
}
\value{
Character vector with the Linear splice junctions file path(s)
}
\description{
circSample and circExperiment class accessors for showing the Linear splice junctions file path(s) of circSample object(s)
}
\examples{
# The Linear splice junctions file path of a sample object
sj.file(sample.object)

# The Linear splice junctions file path in each sample of an experiment
sj.file(experiment.object)

}
