% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{SJTrack}
\alias{SJTrack}
\title{Create a track of linear splice junction}
\usage{
SJTrack(sj = NULL)
}
\arguments{
\item{sj}{Data on linear splice sites. Should be a GRanges object returned by readSJout. Ideally, the data should be restricted to the genomic reange of a single gene (e.g. GenomicFeatures::subsetByOverlaps(sj, region.of.intereset, type = "within")).}
}
\description{
Function to create a CustomTrack of linear splice junctions for plotting using Gviz.
}
