% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{STARlogParser}
\alias{STARlogParser}
\title{Read and parse Log.final.out file}
\usage{
STARlogParser(file, return = "All")
}
\arguments{
\item{file}{Path to file.}

\item{...}{
  Arguments passed on to \code{\link[readr:read_delim]{readr::read_tsv}}
  \describe{
    \item{\code{n_max}}{Maximum number of records to read.}
  }}
}
\value{
Tibble data frame with data on chimeric reads
}
\description{
Read and parse Log.final.out file
}
\examples{
#tmp <- readChimFile(fn = "Chimeric.out.junction")
}
