% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{summarizeBSJreads}
\alias{summarizeBSJreads}
\title{Summarize number of reads per backsplice junction.}
\usage{
summarizeBSJreads(object, cores = 1L, applyFilter = TRUE, ...)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{cores}{Only supported on unix! Number of cores used for parallel processing for circExperiment objets.}

\item{applyFilter}{Boolean indicating whether to applyFilter (default is TRUE).}
}
\value{
Object with same class as input, but with bsj.count slot populated with data
}
\description{
Summarize number of reads per backsplice junction.
}
\examples{
summarizeBSJreads(object)

}
