% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_data_manipulation.r
\name{swapStrands}
\alias{swapStrands}
\title{Reverse strand information and donor/acceptor order, in the chimeric data}
\usage{
swapStrands(data)
}
\value{
The input data where the strand information, and the donor/acceptor order has been reversed.
}
\description{
Reverse strand information and donor/acceptor order, in the chimeric data
}
