% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R
\name{vizJunctions}
\alias{vizJunctions}
\title{Visualize the linear and backsplice junction of a specific gene or genomic region in a specific sample.}
\usage{
vizJunctions(
  object,
  symbol = NULL,
  range = NULL,
  db,
  xlim = NULL,
  onlyJunctionsWithinRange = FALSE,
  validExonModels = NULL,
  deducedCoverage = NULL,
  chromosomes = TRUE,
  path = NULL,
  prefix = NULL,
  device = "png",
  cores = 1L,
  ...
)
}
\arguments{
\item{object}{circSample}

\item{symbol}{Character vector of length 1 with the symbol (genename) of the gene to plot.}

\item{sampleIndex}{Relevant if we should support circExperiment. Numeric value corresponding to the sample to plot.}

\item{sampleName}{Relevant if we should support circExperiment. Name of the sample to plot.}

\item{r}{GenomicRange corresponding to the region to plot.}
}
\value{
A plot
}
\description{
Visualize the linear and backsplice junction of a specific gene or genomic region in a specific sample.
}
\examples{
#vizJunctions(testSample, symbol = "FIRRE")
}
