% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_data_manipulation.r
\name{wrapperFunction}
\alias{wrapperFunction}
\title{Process Chimeric.out.junction file(s) to identify and quantify backsplice sites}
\usage{
wrapperFunction(path, stranded = T, ...)
}
\arguments{
\item{path}{A character vector (length 1 or more) with paths to Chimeric.out.junction files.}
}
\description{
Process one Chimeric.out.junction file to produce files with information on all reads covering candidate backsplice sites and quantitative summary.
}
