% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_generics.R, R/circulaR.r
\docType{package}
\name{circulaR}
\alias{circulaR}
\title{circulaR pipeline wrapper}
\usage{
circulaR(object, annotationDB, cores = 1L, ...)
}
\arguments{
\item{object}{circSample or circExperiment object}

\item{annotationDB}{A TxDb or EnsDb object}

\item{cores}{Only supported on unix! Number of cores used for parallel processing for circExperiment objets.}

\item{...}{}

\item{chromosomes}{Character vector on which chromosomes to keep, while discardin all other chromosomes. The default value (NULL)}
}
\value{
circSample or circExperiment with all data slots populated with processed data
}
\description{
Performs all nescesary steps for filtering, identifying and generate information on backsplice junction candidates

The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{circulaR functions}{

The foo functions ...
}

\examples{
circulaR(circExperiment)

}
