\name{JonesPewsey}
\alias{djonespewsey}
\alias{jonespewsey}

\title{Jones \& Pewsey Density Function}
\description{
  Density for the Jones \& Pewsey circular distribution. 
}

\usage{djonespewsey(x, mu, kappa, psi)}

\arguments{
  \item{x}{a vector. The object is coerced to class \code{\link{circular}}.}
  \item{mu}{direction parameter of the distribution. The object is coerced to class \code{\link{circular}}.}
  \item{kappa}{non-negative concentration parameter of the distribution.}
  \item{psi}{real shape parameter.}
}

\details{}

\value{The density}

\references{Jones , M.C. & Pewsey, A. (2005). A family of simmetric distributions on the circle. J. Am. Statist. Assoc. 100,
   1422–1428}
\author{Federico Rotolo}

\keyword{distribution}
\keyword{circle}
\keyword{circular}
